.ipynb_checkpoints/requirements-checkpoint.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
absl-py==2.3.1
    # via tensorboard
click==8.2.1
    # via clone-oscillations-qat (pyproject.toml)
colorama==0.4.6
    # via click
filelock==3.19.1
    # via torch
fsspec==2025.9.0
    # via torch
grpcio==1.74.0
    # via tensorboard
jinja2==3.1.6
    # via torch
markdown==3.8.2
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.2
    # via
    #   clone-oscillations-qat (pyproject.toml)
    #   scipy
    #   tensorboard
    #   torchvision
packaging==25.0
    # via
    #   pytorch-ignite
    #   tensorboard
pillow==11.3.0
    # via
    #   clone-oscillations-qat (pyproject.toml)
    #   tensorboard
    #   torchvision
protobuf==6.32.0
    # via tensorboard
pytorch-ignite==0.5.2
    # via clone-oscillations-qat (pyproject.toml)
ruff==0.12.11
    # via clone-oscillations-qat (pyproject.toml)
scipy==1.16.1
    # via clone-oscillations-qat (pyproject.toml)
setuptools==80.9.0
    # via
    #   tensorboard
    #   torch
sympy==1.14.0
    # via torch
tensorboard==2.20.0
    # via clone-oscillations-qat (pyproject.toml)
tensorboard-data-server==0.7.2
    # via tensorboard
timm==0.4.12
    # via clone-oscillations-qat (pyproject.toml)
torch==2.8.0+cu126
    # via
    #   clone-oscillations-qat (pyproject.toml)
    #   pytorch-ignite
    #   timm
    #   torchvision
torchvision==0.23.0+cu126
    # via
    #   clone-oscillations-qat (pyproject.toml)
    #   timm
typing-extensions==4.15.0
    # via torch
werkzeug==3.1.3
    # via tensorboard
